// src/model/TreeDetail.ts
export interface TreeDetail {
  oldtreeinfoId: any;
  age: any;
  status: any;
  height: any;
  diameter: any;
  avgcrown: any;
  sncrown: any;
  ewcrown: any;
  recordingtime: any;
  blorg: string;
  number: any;
  cname: string;
  nickname: string;
  lname: string;
  sectionname: string;
  genusname: string;
  address: any;
  level: any;
  type: string;
  altitude: any;
  createtime: any;
  orgId: number;
  longitude: string;
  latitude: string
}